Replaces /refactor -> /lib
