+++ lib/fireferret.js
@@ -229,7 +229,7 @@ class FireFerretClient {
       if (partialFulfillment) return loadFromDB()
 
       this.cache.set(qK, loadOp)
-      log('🔥🧙 Found cached match!')
+      log(`🔥🧙 Found cached superset! Using query '${match}'`)
 
       return docs
     }
@@ -277,8 +277,6 @@ class FireFerretClient {
         }
       }
     }
-
-    log(`Match? ${match}`)
 
     // no match found; read-through to DB.
     if (!match) return loadFromDB()
