+++ lib/util/hash.js
